Fixes test case
Since `article.less` didn’t have an underscore, you were trying to
compile it to `article.css` as well. This is the file that was causing
the error; another way to fix it would have been to import mixins into
both. LESS doesn’t impose this convention as strongly as Sass or
Stylus, but it’s usually best if you name all your partials with an
underscore at the beginning or put them in an underscore’d directory.